refactor: ui controller 와 api controller 분리
