+++ src/main/java/cart/controller/api/ProductController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.api;
 
 import cart.controller.dto.request.ProductRequest;
 import cart.controller.dto.response.ProductResponse;
@@ -19,11 +19,11 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/products")
-public class ProductApiController {
+public class ProductController {
 
     private final ProductService productService;
 
-    public ProductApiController(ProductService productService) {
+    public ProductController(ProductService productService) {
         this.productService = productService;
     }
